refactor!: rename plugin -> positions
