+++ src/apps/moola/positions.e2e.ts
@@ -1,8 +1,8 @@
-import plugin from './plugin'
+import hook from './positions'
 
 describe('getPositionDefinitions', () => {
   it('should get the address definitions successfully', async () => {
-    const positions = await plugin.getPositionDefinitions(
+    const positions = await hook.getPositionDefinitions(
       'celo',
       '0x2b8441ef13333ffa955c9ea5ab5b3692da95260d',
     )
