+++ src/apps/ubeswap/positions.e2e.ts
@@ -1,8 +1,8 @@
-import plugin from './plugin'
+import hook from './positions'
 
 describe('getPositionDefinitions', () => {
   it('should get the address definitions successfully', async () => {
-    const positions = await plugin.getPositionDefinitions(
+    const positions = await hook.getPositionDefinitions(
       'celo',
       '0x2b8441ef13333ffa955c9ea5ab5b3692da95260d',
     )
@@ -11,7 +11,7 @@ describe('getPositionDefinitions', () => {
   })
 
   it('should get no definitions for an address with no blockchain interaction', async () => {
-    const positions = await plugin.getPositionDefinitions(
+    const positions = await hook.getPositionDefinitions(
       'celo',
       '0x0000000000000000000000000000000000007E57',
     )
